third_party/WebKit/LayoutTests/fast/dom/document-preferredStylesheetSet-use-counter.html

<!DOCTYPE html>
<title>document preferredStylesheetSet UseCounter</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
</style>
<script>
test(function() {
  const DocumentGetPreferredStylesheetSet = 2211; // From enums.xml
  assert_false(internals.isUseCounted(document, DocumentGetPreferredStylesheetSet));
  document.preferredStylesheetSet;
  assert_true(internals.isUseCounted(document, DocumentGetPreferredStylesheetSet));
}, 'document.preferredStylesheetSet is use counted');
</script>
